pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uncertain_worms"
version = "0.1.0"
description = "A description of your project"
# You may add other metadata like authors, URLs, etc., if needed

# Dependencies
dependencies = [
    "numpy==1.26.4",
    "gymnasium==0.28.1",
    "minigrid==2.3.1",
    "matplotlib>=3.9.2",
    "imageio>=2.35.1",
    "black==23.9.1",
    "docformatter==1.7.5",
    "isort==5.12.0",
    "python-dotenv==1.0.1",
    "openai==1.53.0",
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
    "pygame==2.6.1",
    "torch==2.2.2",
    "tensorboard==2.18.0",
    "pyro-ppl==1.9.1",
    "graphviz==0.20.3",
    "mypy==1.14.1",
    "pre-commit==4.0.1",
    "requests==2.32.3",
    "pytest>=7.0.0",
    "open3d>=0.19.0",
    "rerun_sdk==0.22.0",
    "pybullet==3.2.7",
    "bosdyn-client==4.1.1",
    "bosdyn-choreography-client==4.1.1",
    "bosdyn-mission==4.1.1",
    "pymdptoolbox @ git+https://github.com/aidan-curtis/pymdptoolbox.git#egg=pymdptoolbox",
    "pyvis==0.3.2",
    "torchvision==0.17.2",
    "opencv-python==4.11.0.86",
    "zmq==0.0.0",
    "polyform@git+https://github.com/williamshen-nz/polyform.git",
    "rerun_robotics@git+https://github.com/aidan-curtis/rerun-robotics.git"
]

# Optional scripts or entry points (if needed)
# [project.scripts]
# example_script = "uncertain_worms:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["uncertain_worms"]

[tool.setuptools]
include-package-data = true